fix a few details, make responsive
