+++ src/pages/activities/Activities.jsx
@@ -1,18 +1,12 @@
 import React from "react";
-import "./activities.css";
-import Navbar from "../../Components/Navbar";
-import Dots from "../../Components/DotPattern";
-import Footer from "../../Components/Footer";
-import { EventList } from "../../Components/EventList";
-import { EventProvider } from "../../contexts/EventContext";
+import Navbar from "../../components/Navbar";
+import Dots from "../../components/DotPattern";
+import Footer from "../../components/Footer";
+import { EventList } from "../../components/EventList.jsx";
+import { EventProvider } from "../../context/EventContext.jsx";
+
 function Header() {
-  return (
-    <div className="flex justify-center items-center">
-      <div className="w-3/4 p-2 flex justify-between items-center">
-        <h1 className="text-white font-semibold text-5xl">Actividades</h1>
-      </div>
-    </div>
-  );
+  return <h1 className="text-white font-semibold text-5xl">Actividades</h1>;
 }
 
 const Activities = () => {
@@ -22,11 +16,10 @@ const Activities = () => {
       <div className="Contenedor w-10/12 mx-auto">
         <Navbar item="eventos" />
         <Header />
-        <div className="container mx-auto mt-8 w-[85%]">
-            <EventProvider>
-              <EventList />
-            </EventProvider>
-          
+        <div className="mx-auto mt-8">
+          <EventProvider>
+            <EventList />
+          </EventProvider>
         </div>
         <Footer />
       </div>
