+++ tailwind.config.cjs
@@ -5,7 +5,6 @@ module.exports = {
     extend: {
       colors: {
         taws: "#1A1E2C",
-        white: "#FFFFFF",
         tawsBlue: "#1AC8FF",
         orange: "#D19121",
         cargo: "#CBCA6D",
